Fix ingestor Status logging, split collector and ingestor Status lines
- Remove dead ingestor samples-stored logging and coordinator write
  path. Collector only sends WAL files to collector now - it does not
  send prom remote-write messages any longer.
- Split the status logging paths for collector and ingestor so that
  ingestor can only log metrics/telemetry relevant to ingestor.
